Updating Clipper tutorial
